Some type movement to make linter happy
