+++ packages/common/testUtil/extractTargetedMarks.ts
@@ -1,40 +1,6 @@
-import { getKey, splitKey } from "..";
-import type { ReadOnlyHatMap } from "../../cursorless-engine/core/IndividualHatMap";
-import type {
-  PrimitiveTargetDescriptor,
-  TargetDescriptor,
-} from "../../cursorless-engine/typings/TargetDescriptor";
-import type { Token } from "../../cursorless-engine/typings/Types";
-
-function extractPrimitiveTargetKeys(...targets: PrimitiveTargetDescriptor[]) {
-  const keys: string[] = [];
-  targets.forEach((target) => {
-    if (target.mark.type === "decoratedSymbol") {
-      const { character, symbolColor } = target.mark;
-      keys.push(getKey(symbolColor, character));
-    }
-  });
-  return keys;
-}
-
-export function extractTargetKeys(target: TargetDescriptor): string[] {
-  switch (target.type) {
-    case "primitive":
-      return extractPrimitiveTargetKeys(target);
-
-    case "list":
-      return target.elements.map(extractTargetKeys).flat();
-
-    case "range":
-      return [
-        ...extractTargetKeys(target.anchor),
-        ...extractPrimitiveTargetKeys(target.active),
-      ];
-
-    default:
-      return [];
-  }
-}
+import { splitKey } from "..";
+import { ReadOnlyHatMap } from "../types/ReadOnlyHatMap";
+import { Token } from "../types/Token";
 
 export function extractTargetedMarks(
   targetKeys: string[],
