+++ packages/common/types/.eslintrc.json
@@ -1,0 +1,25 @@
+{
+  "rules": {
+    "@typescript-eslint/no-restricted-imports": [
+      "error",
+      {
+        "patterns": [
+          {
+            "group": ["@cursorless/*", "../../*"],
+            "message": "Common shouldn't have any dependencies"
+          }
+        ],
+        "paths": [
+          {
+            "name": "vscode",
+            "message": "Common shouldn't have any dependencies"
+          },
+          {
+            "name": "@cursorless/vscode-common",
+            "message": "Common shouldn't have any dependencies"
+          }
+        ]
+      }
+    ]
+  }
+}
