+++ packages/common/types/RangeOffsets.ts
@@ -1,0 +1,4 @@
+export interface RangeOffsets {
+  start: number;
+  end: number;
+}
