+++ packages/cursorless-engine/core/updateSelections/getOffsetsForNonEmptyRangeInsert.ts
@@ -3,8 +3,8 @@ import { leftAnchored, rightAnchored } from "../../util/regex";
 import {
   ChangeEventInfo,
   FullRangeInfo,
-  RangeOffsets,
 } from "../../typings/updateSelections";
+import { RangeOffsets } from "@cursorless/common";
 
 /**
  * Gets updated offsets for the range `rangeInfo` after the change described by
