+++ packages/cursorless-engine/core/updateSelections/getUpdatedText.ts
@@ -1,8 +1,8 @@
 import {
   ChangeEventInfo,
   FullRangeInfo,
-  RangeOffsets,
 } from "../../typings/updateSelections";
+import { RangeOffsets } from "@cursorless/common";
 
 /**
  * Updates the text of the given rangeInfo to take into account the given change.
